[l10n] Only show the InfoBar when the locale differs from en-US
This CL fixes a bug reproduced the following way:
  1) User enables DevTools locale experiment
  2) User switches to zh locale and restarts DevTools
  3) The InfoBar is shown with the notification that DevTools is now
     available in English

Instead of having yet another hidden setting that stores whether the
user has touched the language setting, we check if the language
setting doesn't have the default value (en-US) and skip showing
the InfoBar. The user already knows about the language setting and
has actively selected a different locale in this case.

R=tvanderlippe@chromium.org

Bug: 1163928
Change-Id: Ifec3a0aeda9c1a3801af1e93ae8201a852bc9ea7