[l10n] Adopt Lighthouse mapping of locales to filenames
This CL adopts the LH appraoch that maps concrete locales to a
translation file. Given that we already use the same list of
translation it makes sense to also adopt the same mapping. LH folks
put a lot of effort in curating and matching this mapping to the ISO
spec.

It works on the "LOCALES" object, where initially the filepath is
written as the value for each locale key. When the actual data was
fetched, we replace the key with the actual translations. Until now,
we only had dummy placeholder values as values.

The current approach is not ideal, but should get us up and running
until we refactor or rewrite the usage of the LH library. Ideally
we have a single source of truth, where we specify for any locale:
the filename, whether its remote/local, whether its a pseudo locale.

Drive-by: Use `module.exports` instead of ES exports. Whatever rollup
is doing breaks the usage of the LOCALES object.

R=tvanderlippe@chromium.org

Bug: 1163928
Change-Id: I0f52dc514710c663fbc8cb73d2171eabf9eb9bea