Cleanup path handling in collect-strings
While reading through this script I spotted a lot of duplication
of `front_end` in its path handling. Since we assume that all
strings are collected from `front_end/`, we can pass in that
path and use that, cleaning up most of the path concatenation
in the script.

DISABLE_THIRD_PARTY_CHECK=Path handling fixes in collect-strings
R=szuend@chromium.org

Bug: none
Change-Id: I0e7d004ad4d1d16f92cd88f3fed94dfe4c4108b0