Reland: Add component for linking to network requests
This CL adds a component that can be used to link to a network request.
The component optionally uses a RequestResolver to resolve the request
if only a requestId (in the form of an affected request) is provided.
The component updates its state once the request becomes available,
thereby supporting plasticity of the UI.

This is a plain reland of https://crrev.com/c/2960617.

Bug: chromium:1217975
Change-Id: Ideccf4f9e64cf451a8c28583a9d72e717ff6706f
Cq-Include-Trybots: luci.devtools-frontend.try:devtools_frontend_mac_rel