Use Map for Material shade colors
We are now mangling any property field that starts with `#`. Since
hex colors use this notation, they were incorrectly mangled.
However, we should use a Map for this instead, since that is
a better data structure and circumvents the mangling problem.

I have also confirmed that there are no other occurrences of
properties starting with a `#` in the codebase.

R=aerotwist@chromium.org

Fixed: 1223298
Change-Id: I8bc5e023babb7fa398995dfc1a82a7c53a398e9a