[Debugger] Properly await Debugger.enable response
Previously we would rely on the back-end handling incoming CDP
requests in order. Largely, this also happens, but some exceptions
are methods such as setBreakpointByURL, which are defined in
devtools_session.cpp. These exceptions may bypass the order of
incoming requests.

As a result, we didn't await the Debugger.enable response
before actually trying to set a breakpoint. This CL waits until
we receive the response before actually setting the internal
flag to be enabled.

Also-by: bmeurer@chromium.org

Bug: b/190756091
Change-Id: I67b659bd6e82e8c54513fb444ddd62f1dad122dd