Add accessibility alert when adding a row in datagrid in devtools
Previously, when user adds a new row to a datagrid, for example, adding
a row to Local Storage in Applications panel, the voice over announces
the content of the row, namely, column names followed by column values
while column values are empty. There is no indication that a new row is
added and there is no indication what can be done next.

We add an aria alert to let the user know that an empty row is added and
that they may edit values of the cells on this row.

Bug: 1223875
Change-Id: I3a0c6d53d60aaa34cc910d3b48bada106ed8b156