Fixes outlines in network overview and color-code setting in network panel
In network overview page, a deprecated CSS variable "--neutral-layer-l4" was being used to draw outlines. Because the CSS variable was missing, it was drawing black lines. Updated the CSS variable color to "--color-background-opacity-80" from theme colors.

Also color-code setting is assigning a fill color instead of CSS color variable which _drawLayers was previously using to set the fill color. Updated the function so that when the setting is toggled on, instead directly use the hex colors instead of using CSS color variable to look up color.

https://imgur.com/a/wKZDDlz


Bug:1216849,1213595
Change-Id: Ie61ee3198873583415ad992cd5b80ce16479f82b