Remove margin from Settings screen
The settings screen has a margin that makes its content 3 pixels smaller
than the DevTools UI. As a result a small gap is seen around its edges
where the DevTools UI can be seen and interacted with. This means that
users can accidentaly interact with other parts of DevTools which are
visible when the Settings screen is open (for instance, closing DevTools
when trying to close the settings screen because of the closeness of
both close buttons). This CL removes such margin so that the settings
screen completely overlays DevTools.

Fixed: 1142747
Change-Id: I26d6c41bcb4ba0d5207911c86fdf4d827b5c9713