Fix tabbed background contrast and close icon bg color
My initial CL (https://crrev.com/c/3000817) fixed the initial bug
report but caused an issue with the close icon. The initial CL also
changed the lightness of the toolbar more than I wanted, so in this CL
I've undone that change and instead provided an override for dark mode
to ensure the active tab is still easily visible.

Fixed: 1226497, 1225549
Change-Id: I9621bb0b6f1d709fa6e68f33c0071df2b0e647eb