[CSS Migration] migrate TrustTokensView
There were two parts to this file:

1. Create `trustTokensView.css` which houses the main styles.
2. The delete button needs to be its own stylesheet due to the fact
   that it gets rendered from within the data grid (e.g. in a
   different shadow DOM) so I pulled that out into its own component.

I also discovered the ESLint component naming rule breaks with
multiple components in one file which I'd like to fix, so I've
disabled the rule and filed a bug for that.

Additionally there was no component doc for the TrustTokensView
component so I added one.

Bug: 1106746, 1226741
Change-Id: Ia1e79adff77c831adec5a28331f9019176ef9977