[wasm] Suppress popover preview for $label identifiers.
Right now we don't have anything meaningful to show in the popover
preview for $label identifiers, and Debug-Evaluate in V8 will just
return `undefined` for them, which is confusing to developers. So
detect these special identifiers (operands to specific control
instructions) and suppress the popover preview for now.

Fixed: chromium:1155548
Change-Id: Ic43910dc12b535bbf572fbdde3c658b760bcb468