Ensure consistent usage of Protocol.Network.RequestId
This uncovered the fact that NetworkRequests really have two
request ids: One that is a string used in the front-end that
doesn't always have a corresponding request in the back-end,
and one that is always a back-end id. This CL makes this
explit in the types.

DISABLE_THIRD_PARTY_CHECK=disable

Bug: chromium:1226471
Change-Id: I894d6ab4d14099b9793c5e1c1a2863d2ec9c2548