Invoke late initialization from meta file
Currently whats new panel is not showing up when version is updated. This is because Late Registration happens through HelpImpl.ts which is not loaded until help view is loaded. This CL updates late registration through -meta file, this way we check if the whats new has to be displayed even if view is not loaded.

Bug:1227244
Change-Id: I9d286dbc8ea2defca9ea3cd4bb6336ffb1d11ea9