[cleanup] Refactor and simplify BreakpointManager a bit.
Some of the code in BreakpointManager.ts was still threading through a
callback despite already using promises and async functions, which allow
for proper chaining.

Also the _addResolvedLocation() method was returning a boolean, but that
was essentially ignored (or worse, led to an inconsistent state).

And last, the breakpoint listener were passed as newly allocated
closures, both upon addition and removal, but since the event listener
machinery uses closure identity, this was clearly not working properly
(for removal). Probably not a serious bug currently, but definitely
unnecessarily surprising.

Bug: chromium:700516
Change-Id: Ic006c8e14e66d036658769106ca34282aec74c20