Introduce `NoUnion<T>` type marker and use it for event dispatch
This CL adds some advanced TypeScript generics to further constrain
the `EventTarget` interface. This will prevent dispatching events
with the wrong payload type when the exact event type is not known
to TypeScript. Example:

enum Event {
  Event1, // takes a number
  Event2, // takes a string
}

someEmitter.dispatchEventToListeners<Event.Event1 | Event.Event2>(
    Event.Event2, 42.0);

Without this CL, the code would compile, because the payload is
typed as `string | number`. With this CL the code won't compile
as the first argument (`eventType`) will be typed as `never` now.

R=sigurds@chromium.org, tvanderlippe@chromium.org

Bug: chromium:1228674
Change-Id: Iac9fdc0a8766104a26c30cb5f8dfaa82ec0e5f89