Fix positioning of palette picker slide-in
The palette picker now slides in from the bottom only as far as
necessary (100% of its size). A few height values are adjusted by a few
pixels, so that the awkward case of seemingly having a grey border
between color spectrum and palette picker when the previously selected
palette has exactly 3 rows ()is avoided.

This reverts the changes of https://crrev.com/c/2947257

Fixed: 1224269, 1216355
Change-Id: Ibe6a6cbfa68e9db5c5535f27c61cefb7215021af