[Debugger] Wait until the Debugger is ready to pause
Previously we waited until the Debugger is enabled before trying to
set a breakpoint for a new target. This CL changes this to wait until
the Debugger is ready to pause, since at that point we only know
that the `Debugger.enable` response has actually been received.

If we set breakpoints too early (before the back-end is ready), we
receive a ServerError and remove that breakpoint on all targets,
which leads to "breakpoints being flakily removed".

This is a quickfix to unblock users. The reason for why we do
not just correctly wait for the `Debugger.enable` response
before internally setting `debuggerEnabled` to be true, is
that other races become more apparent, which has to be
investigated further.

Also-by: bmeurer@chromium.org
Bug: b/190756091, chromium:1229541
Change-Id: If141848f5e6c4a064d324ecde2b2795b99a5406e