Fix viewport size in headful mode
In headful mode, the viewport size was different from the
headless viewport because Puppeteer's default viewport is
800x600 leading to some test failues.

This CL makes sure that the same viewport is set in headful mode
and that the window size is somewhat bigger than the viewport.
It's hard to know what window size should be and that probably
varies from platform to platform so setting some best guess values
that make the window bigger than the viewport on glinux.

Bug: none
Change-Id: I44ff9ac0758a9e79d771b58b38a0e09cbe9e8785