[DataGrid] fix extra padding on filler row
The data-grid has a filler row that it renders to fill any additional
height of the container, but when the data-grid was filling the height
anyway this row would be 2px high (see screenshot in bug).

This CL changes the row to be 0px high, but I then found an oddness
with the rendering engine (not clear to me if it's desired behaviour
or a bug) where if the table had no rows (so we want the padding row
to be 100% high to fill the space) but then the padding row was given
`padding: 0`, it would not take up any height. Therefore we
conditionally apply an extra CSS class to give the row some padding
such that it takes up extra height in the event of the table being
empty.

Fixed: 1226745
Change-Id: I9030832320eb2554adbac1dc92dfbd8ec4cf9c2a