Created ESLint rule and tests to migrate legacy files
It adds `import componentStyles from './component.css.js'` for any
registerRequiredCSS('front_end/component/component.css') call.

It also checks if there is an existing `wasShown()` method. If this
exists then it adds an extra line to add the
`this.adoptedStyleSheets = [componentStyles]` statement. Otherwise,
the `wasShown()` method is created and added.

Bug: 1106746
Change-Id: I7d0244d129fe0f4a50432fd7b53135760ae304ae