[l10n] Do not try to translate experiment names
Experiment titles do not have translations yet so attempting to
retrieve a translated title is a no-op. The upcoming stricter i18n
library will throw in this case instead of silently ignoring the
wrong API usage.

Follow-up work will introduce proper UIString entries for
experiments that can actually be translated, as well as change the
type of `Experiment#title` to be a localized string.

R=sigurds@chromium.org

Bug: 1232313
Change-Id: I96cca6171e925eff99bb78040fe0cf07cfb07c1d