[i18n] Upgrade intl-messageformat to 9.8.0
This CL upgrades the intl-messageformat package to version 9.8.0. This
is the first version that includes a modern ESM web bundle of the
library with all dependencies included.

Beware: After this CL lands, the i18n-bundle bundling process while be
temporarily broken, until third_party/i18n/localized-string-set is
switched over to use IntlMessageFormat directly from the package
instead of via the i18-bundle.

R=tvanderlippe@chromium.org

Bug: 1231873
Change-Id: Ia0d602f780491a05034e68f3f28e91d2b184d631