[console] Associate `queryObjects` console message with correct context.
The `ConsoleMessage` that was created for `queryObjects(obj)` did not
have an execution context associated with it, which meant that it'd not
show at all as soon as the "Selected context only" setting in the
Console was enabled. This uses the `executionContextId` reported by the
`Runtime.inspectRequested` as of https://crrev.com/c/3046187 to properly
associate the `ConsoleMessage` with the execution context in which the
`queryObjects(obj)` expression was originally executed.

Drive-by-fix: Migrate the `QueryObjectRequested` event handler to the
new type safe event listeners mechanism.

Bug: chromium:1228674
Change-Id: I12328c6ebcca2aea38b14bd8f9d50ec76014073c
Fixed: chromium:1231521