[i18n] Add entrypoint for intl-messageformat
This CL adds a BUILD.gn and devtools_entrypoint to intl-messageformat.
DevTools modules can then import intl-messageformat via the standard
mechanism.

After this CL lands, the TS files in third_party/i18n can be migrated
to import from the entrypoint rather than the IntlMessageFormat
version that is part of the i18n-bundle. After that migration, the
i18n-bundle can also be simplified and no longer needs to include
IntlMessageFormat.

R=tvanderlippe@chromium.org

Bug: 1231873
Change-Id: Ie74378bf76211669fe088670ff2bfa8c970a2d3c