Added edge case support for migrating legacy calls to `registerRequiredCSS`.
There are two main edge cases: multiple calls to
`this.registerRequiredCSS` and calls to properties of the object, e.g.
`this._widget.registerRequiredCSS`. I have added cases to the ESLint
rule to migrate these automatically.

In ESLint tests, the output is produced after only a single pass of
the rule. Keeping this in mind, the tests for multiple calls have
been broken down into multiple tests cases.

Bug: 1106746
Change-Id: If5a4827ec7230508c9208c6d23882499495be12a