[i18n] Manually add types for @formatjs/icu-messageformat-parser
This CL copies the .d.ts files from icu-messageformat-parser into
third_party/intl-messageformat. Otherwise, the underlying AST of the
IntlMessageFormat class will be typed as any, making it easy to
introduce mistakes.

Currently, the only user of these types is `getFormatLocalizedString`
which allows HTML elements as placeholders and manually walks the
AST to stich together the resulting parent HTML element.

R=tvanderlippe@chromium.org

Bug: 1231873
Change-Id: Ib964632ebfef4bd87c44e9fd75a2843f55d51ce1