Prioritize HTTP errors over cors errors in network status column
Previously, DevTools would mark 404s on CORS requests as cors errors,
and point out missing headers. While the headers are *technically*
missing, the information that the request 404'd is more important to
web developers.

This CL prioritzes HTTP error codes (>=400) over CORS errors in the
network panel's status column. This was previously not possible, as
for CORS errors the underlying reponse code was not reported, which
changed in https://crrev.com/c/3043654.

Bug: chromium:782442
Change-Id: I38d9acc938bfbfe3b8f63e49cf81aaf7f0429fbe