Updated scripts to resolve paths for css files not in same directory and fixed duplicates in GRD files
Some files import CSS files not in the same directory as them. The
script was currently unable to handle this case and so migrations led
to `File Not Found` errors for the `.css.js` files. The script now
finds the relative path from the imported file to the current file and
adapts the import statement correctly.

Duplicates were being added to `devtools_grd_files` as the `includes`
check was not using the entire file path. This also led to errors
during the build.

Bug: 1106746
Change-Id: I5da808a8885bc18477cfa31801a904e17924caab