Support navigating iframes in accessibility tree
This change also introduces e2e test suites for both the new
accessibility tree and the existing accessibility pane.
The fuller accessibility tree now listens for documentUpdated events
to refresh the root nodes. This is a temporary work-around to fix a bug
where the tree would be empty if we attempt to poppulate it before the
DOM is ready. More fine-grained updates of the tree will be tackled in
a follow-up.

Bug: 1226486
Change-Id: I61edd34724da192225b31348bf9300df212fb11a