[issues] Add affected resource view for raw cookie lines
This CL adds a very simple view for Issues that don't have a canonical
cookie but only a raw `Set-Cookie` header line. Unfortunately, there is
nothing we can link the raw cookie line to. There is no canonical
cookie in the cookies panel and a network request is also not available
in all cases (e.g. for cookies set via document.cookie).

R=sigurds@chromium.org

Bug: 1179186
Change-Id: If502f4eeb1809378db9f8363ba74a231d01c6063