Move sensors panel into separate folder
Previously, this code was living in panels/emulation. However,
there is code relation between sensors and emulation, other than
they are both concerning emulation techniques.

Since the panels/emulation code is run in a separate frame for
device mode emulation, we need to limit the amount of code to run,
to unblock using adoptedStyleSheets for our style sheets. Therefore,
if we move this code out, we have to deal with fewer problems, while
it also makes more sense to put the code in a separate folder.

R=jacktfranklin@chromium.org,mathias@chromium.org
CC=kritisapra@google.com

Bug: 1106746
Change-Id: I788550256d70c64027e584c795d08177d69a0984