Move emulation model into models/
This code concerned a model implementation that should instead live
in models/. Initially I thought that I was able to remove all dependencies
on ui/, but I found out that `UI.Geometry.Size` was still used.

In a follow-up CL, I will move that class into `core/platform`, since
it is not UI-concerned and will remove the dependency on `ui/` from
a fair bit of model code again.

R=jacktfranklin@chromium.org

Bug: 1234798
Change-Id: Ic32b11a82392a22022bafaf36b0a959d4ae9f9ad