Introduce compact mode for error-warning counter
This CL adds a compact mode to the error-warning counter in
DevTools top-right corner. The compact mode is activated when
the DevTools window's width becomes too small show show at
least the chevron of the main tab bar. It deactivates when
there is twice as much space available as the toolbar in compact
mode needs.

This CL also adjusts the minimum width of the DevTools window
to 250px (up from 240px).

Screenshot: https://imgur.com/a/DK06j9m

Bug: chromium:1232370
Change-Id: I842b514089e255adb09165497db053c3689d43b8