[issues] Link raw cookie lines to network panel
This CL adds a new network filter that allows users to search for
requests where the response includes a specified value in the
`Set-Cookie` header.

The issues panel uses the new filter to link issues with raw
cookie lines back to the network panel (should the issue be
associated with a network request).

Screenshots: https://imgur.com/a/zVZNrK1

R=sigurds@chromium.org

Bug: 1179186
Change-Id: I9e275d31676a67a1b197c9aa195b65a144caeebf