Change selected animation row background color to be transparent
The transparency was lost during the dark mode migration and when a row
is selected the scrubber is chopped as a result. This replaces the
background color of the selected row to a variable which is partially
transparent.

Before: https://imgur.com/a/uc4Xm7e
After: https://imgur.com/a/6l6sWRh

Fixed: 1235961
Change-Id: I18aa4f4e40def5220c7b3caf687041656d51e74a