Fix filename in generated tsconfig
The generated JavaScript file ends with `.js` and therefore we
should list the full name in our tsconfig as well. We ran into
compilation issues on https://crrev.com/c/3073941 and they required
the full extension to be listed. It's not clear to me why we
didn't run into this before, but it doesn't seem to break
any of the existing CSS file inclusions either.

R=jacktfranklin@chromium.org
CC=alexrudenko@chromium.org,kritisapra@google.com

Bug: 1106746
Change-Id: I45d593f4aee56b783349c82ea91cbde9b2db7d66