[Underscore Migration] Migrate panels/profiler
This patch is part of the underscore migration that adds the private
keyword to all properties that start with an underscore and removes the
underscore.

Bug:chromium:1207342


[Underscore Migration] Migrate panels/profiler

This patch is part of the underscore migration that adds the private
keyword to all properties that start with an underscore and removes the
underscore.

R=tvanderlippe@chromium.org

Bug:chromium:1207342
Change-Id: I03e6ee1f2d3a872dff1c0695dac07479131593c2