Add types and unit tests to DockController
This CL adds some types and unit tests to the DockController. I did this as I
was using it and found myself writing types in my code to wrap around, so I
thought I'd come back and try to land a change that improved and added types, along with some basic unit tests (emphasis on "basic"!).

Bug: none
Change-Id: Ic715f003ad69ea1388ecdf5ea0a1c91dad38cd05