Increased var maxHighlightLength value in order to highlight/search longer lines
As mentioned in comment #2 of the bug report, when a non pretty printed line is searched, codeMirror only searches/highlights results within the first 1,000 characters. This is a limitation set on a codeMirrorTextEditor setting “maxHighlightLength”. CodeMirror by default has this setting set to 10,000 which from what I’ve seen still maintains responsiveness and I did not witness any additional lag.

https://drive.google.com/file/d/1f0ePqGOq4FCBdolXyzaHzM-Von6jWpch/view?usp=sharing

R=tvanderlippe@chromium.org

Bug: 1226169
Change-Id: I4176fc0561613b20f59c01b6836495e34e042541