Cached IssueViews hold cleared instances to aggregated issues after Event.FullUpdateRequired is fired in the IssuesManager.
Whenever an Issue is hidden through the hide issues menu, Event.FullUpdateRequired is dispatched in the IssuesManager. This invokes the onFullUpdateRequired method in IssueAggregator, which clears all AggregatedIssue instances both aggregatedIssueByCode and hiddenAggregatedIssueByCode maps. Since we cache IssueViews through the clearViews in the IssuePane. Cached IssueViews do not update count of aggregatedIssues as the instances they hold are no longer present in the IssueAggregator.
Bug:1175722
Change-Id: I934adbdcf410141c46e440ddb06c20c642d56d24