Heuristically identify identifier types in CDP
This CL uses a heuristic to classify identifier types:
If the type ends with Id or ID it is considered an
identifier type. The CL also introduces an override
in protocol_dts_generator.ts which can be used to
add more types, or disable the heuristic for specific
types.

DISABLE_THIRD_PARTY_CHECK=disable

Bug: chromium:1226471
Change-Id: I4696463622ef026c48e1abeefd48aa438f9d11f6